geyser: Move notifier interfaces into new agave-geyser-notifier-interface crate
The Geyser notifier interfaces were defined inside validator internals
(accounts-db, ledger, rpc, gossip, plus a runtime type in the manager's
block metadata interface) and implemented by the plugin manager, forcing
the manager to depend on those crates at the exact validator version.
This dependency direction is the main obstacle to consuming the Geyser
crates from outside the monorepo.

Introduce agave-geyser-notifier-interface, a leaf crate depending only
on stable crates.io types, and move into it:
- AccountsUpdateNotifierInterface + AccountForGeyser (from accounts-db)
- DeshredTransactionNotifier + UpdateParentInfo (from ledger)
- SlotStatusNotifierInterface and SlotNotification (from rpc)
- ContactInfoSnapshot/Event + channel aliases (from gossip)
- BlockMetadataNotifier (from the manager itself)

The defining crates re-export from the new crate at their existing
paths, so no import path breaks. UpdateParentInfo::from_slot_meta
becomes blockstore_meta::update_parent_info_from_slot_meta (an inherent
impl cannot live outside the defining crate), and notify_block_metadata
now takes &[(Pubkey, BlockRewardInfo)] + Option<u64> instead of
runtime's KeyedRewardsAndNumPartitions, converted once per block in
replay_stage.

The plugin manager drops its solana-accounts-db and solana-runtime
dependencies. EntryNotifier (ledger) and TransactionNotifier (rpc) stay
put for now: their signatures contain VersionedBlockFooter and
TransactionStatusMeta, whose type chains are not yet stably published.
