+++ geyser-notifier-interface/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "agave-geyser-notifier-interface"
+description = "Validator-internal notifier interfaces implemented by the Geyser plugin manager."
+documentation = "https://docs.rs/agave-geyser-notifier-interface"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+rust-version = { workspace = true }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+all-features = true
+rustdoc-args = ["--cfg=docsrs"]
+
+[features]
+agave-unstable-api = []
+
+[dependencies]
+crossbeam-channel = { workspace = true }
+solana-account = { workspace = true }
+solana-clock = { workspace = true }
+solana-hash = { workspace = true }
+solana-message = { workspace = true }
+solana-pubkey = { workspace = true }
+solana-reward-info = { workspace = true }
+solana-signature = { workspace = true }
+solana-transaction = { workspace = true }
+
+[lints]
+workspace = true
