+++ ledger/Cargo.toml
@@ -30,6 +30,7 @@ agave-unstable-api = []
 
 [dependencies]
 agave-feature-set = { workspace = true }
+agave-geyser-notifier-interface = { workspace = true }
 agave-reserved-account-keys = { workspace = true }
 agave-snapshots = { workspace = true }
 agave-transaction-view = { workspace = true }
