+++ rpc/Cargo.toml
@@ -31,6 +31,7 @@ dev-context-only-utils = [
 
 [dependencies]
 agave-feature-set = { workspace = true }
+agave-geyser-notifier-interface = { workspace = true }
 agave-snapshots = { workspace = true }
 agave-votor-messages = { workspace = true }
 base64 = { workspace = true }
